@@ -3,28 +3,31 @@
   "homepage": "http://www.jaosimt.com",
   "version": "0.1.0",
   "private": true,
+  "engines": {
+    "node": "<=20"
+  },
   "dependencies": {
-    "@atlaskit/css-reset": "^6.2.0",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "deep-equal": "^2.0.5",
-    "gh-pages": "^3.1.0",
-    "node-sass": "^5.0.0",
-    "react": "^17.0.2",
-    "react-beautiful-dnd": "10.1.1",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.3.3",
-    "web-vitals": "^1.0.1"
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "deep-equal": "^2.2.3",
+    "gh-pages": "^6.1.1",
+    "node-sass": "^9.0.0",
+    "react": "^18.2.0",
+    "react-beautiful-dnd": "13.1.1",
+    "react-dom": "^18.2.0",
+    "react-parse-htmlstring": "^4.0.1",
+    "react-router-dom": "^6.22.3",
+    "react-scripts": "5.0.1",
+    "styled-components": "^6.1.8",
+    "web-vitals": "^3.5.2"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "predeploy": "npm run build",
+    "predeploy": "yarn run build",
     "deploy": "gh-pages -d build"
   },
   "eslintConfig": {
